Fix several problems with lag preserving
(cherry picked from commit 22a12df594089286e3cfb1dbbec31675551e468e)

+++ sources/include/client.h
@@ -83,7 +83,7 @@ struct od_client {
 	char *ldap_auth_dn;
 #endif
 
-	int last_catchup_lag;
+	int64_t last_catchup_lag;
 
 	/* external_id for logging additional info about client authneticated with external auth */
 	char *external_id;
