+++ sources/machinarium/machine.c
@@ -383,7 +383,7 @@ MACHINE_API uint64_t machine_time_us(void)
 	return mm_self->loop.clock.time_us;
 }
 
-MACHINE_API uint32_t machine_timeofday_sec(void)
+MACHINE_API int64_t machine_timeofday_sec(void)
 {
 	mm_clock_update(&mm_self->loop.clock);
 	return mm_self->loop.clock.time_sec;
